backend/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
gunicorn
flask-jwt-extended
werkzeug
flask-cors
bcrypt
Flask-Mail==0.9.1